refactor(verification): route verify activity through engine run_profile (#630 PR2)
Make the declarative engine live: the execute_requirement_verification
Durable activity now calls run_profile instead of run_verification. For
plain ValidatorDescriptors (every phase today) run_profile is behaviorally
identical, so no phase changes behavior.

- Delete run_verification (superseded by run_profile).
- Move execute_verification_job into engine.py (it composes run_profile
  with prepare/persist; belongs at the engine layer and resolves the
  executor->engine import direction cleanly).
- persist_verification_result strips evidence bundles before the terminal
  DB write so file contents never reach Postgres.
- Repoint tests to patch validate_submission where the legacy check imports
  it (verification.engine) and import the moved symbols from engine.

Grading path untouched: the self-gating collect_llm_grading_requests probe
stays until phases declare llm_rubric_review steps, so no grading regression
in intermediate merges.

Copilot-Session: 64b202e1-77f6-431d-89ab-48bb301cc8a3
